Program 1: Ray Tracer
1. Overview

Ray tracing is a simple and powerful algorithm for rendering images. Within the accuracy of the 
scene and shading models and with enough computing time, the images produced by a ray tracer 
can be physically accurate and can appear indistinguishable from real images. Your ray tracer will 
not be able to produce physically accurate images though. 


2. Requirements

1. Use a ray tracing algorithm. 
2. Support spheres. 
3. Support the Lambertian and Blinn-Phong shading models, as defined in Shirley 9.1–9.2 and 
in the lecture notes. 
4. Support point lights that provide illumination that does not fall off with distance. 
You do not need to worry about malformed input, either syntactically or semantically. For instance, 
you will not be given a sphere with a negative radius or a scene without a camera.